# Conflicts:
#	parent/composant-business/document-interface/src/main/java/cmr/notep/modele/Attributs.java
#	parent/composant-business/document-interface/src/main/java/cmr/notep/modele/Distributeurs.java
#	parent/composant-business/document-interface/src/main/java/cmr/notep/modele/Familles.java
#	parent/composant-business/document-interface/src/main/java/cmr/notep/modele/Missions.java
#	parent/composant-business/document-interface/src/main/java/cmr/notep/modele/Mouvements.java
#	parent/composant-business/document-interface/src/main/java/cmr/notep/modele/PersonnesPhysique.java
#	parent/composant-business/document-interface/src/main/java/cmr/notep/modele/Promotions.java
#	parent/composant-business/document-interface/src/main/java/cmr/notep/modele/Ressources.java
#	parent/ittest/src/test/resources/data-h2.sql
#	parent/ittest/src/test/resources/schema-h2.sql
#	parent/ressources-jpa/document-jpa/src/main/java/cmr/notep/dao/AttributsEntity.java
#	parent/ressources-jpa/document-jpa/src/main/java/cmr/notep/dao/DistributeursEntity.java
#	parent/ressources-jpa/document-jpa/src/main/java/cmr/notep/dao/DocEtatsEntity.java
#	parent/ressources-jpa/document-jpa/src/main/java/cmr/notep/dao/MouvementsEntity.java
#	parent/ressources-jpa/document-jpa/src/main/java/cmr/notep/dao/PersonnesEntity.java
#	parent/ressources-jpa/document-jpa/src/main/java/cmr/notep/dao/PromotionsEntity.java

+++ parent/composant-business/document-interface/src/main/java/cmr/notep/modele/PrecoMouvementsQtes.java
@@ -1,5 +1,6 @@
 package cmr.notep.modele;
 
+import com.fasterxml.jackson.annotation.JsonBackReference;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import lombok.*;
 
@@ -16,8 +17,8 @@ public class PrecoMouvementsQtes {
     private String id ;
     private int qteMin ;
     private int qteMax ;
-    private int montantMin ;
-    private int montantMax ;
+    private double montantMin ;
+    private double montantMax ;
     private PrecoMouvements precoMouvement ;
     private List<Familles> familles;
     private Ressources ressource;
